% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ac}
\alias{ac}
\title{Employee costs accounting codes}
\format{An object of class \code{data.frame} with 20 rows and 2 columns.}
\usage{
ac
}
\description{
M3 codes used by accounting for employee costs.
}
\keyword{datasets}
