% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acSAP}
\alias{acSAP}
\title{Cost accounting codes in SAP}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 123 rows and 2 columns.
}
\usage{
acSAP
}
\description{
SAP cost codes used by accounting.
}
\keyword{datasets}
