% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_fun_assign.R
\docType{methods}
\name{assignEmp}
\alias{assignEmp}
\alias{assignEmp,Employee-method}
\alias{assignEmp,Staff-method}
\alias{assignEmp,Non Staff-method}
\alias{assignEmp,Clerk-method}
\alias{assignEmp,Operation Personnel-method}
\alias{assignEmp,Technical-method}
\alias{assignEmp,Production Personnel-method}
\alias{assignEmp,Supervisor-method}
\alias{assignEmp,Laborer-method}
\alias{assignEmp,Operator-method}
\title{Assign an employee}
\usage{
assignEmp(empT, empR)

\S4method{assignEmp}{Employee}(empT, empR)

\S4method{assignEmp}{Staff}(empT, empR)

\S4method{assignEmp}{`Non Staff`}(empT, empR)

\S4method{assignEmp}{Clerk}(empT, empR)

\S4method{assignEmp}{`Operation Personnel`}(empT, empR)

\S4method{assignEmp}{Technical}(empT, empR)

\S4method{assignEmp}{`Production Personnel`}(empT, empR)

\S4method{assignEmp}{Supervisor}(empT, empR)

\S4method{assignEmp}{Laborer}(empT, empR)

\S4method{assignEmp}{Operator}(empT, empR)
}
\arguments{
\item{empT}{an \code{\link{Employee-class}} object representing a theoretical
employee}

\item{empR}{an \code{\link{Employee-class}} object representing a real
employee}
}
\value{
a list containing the following:
  \enumerate{
    \item a \code{\link{data.frame}} containing the man hours database
      resulting from the employee assignment. This is consisted of the
      following columns:
      \describe{
        \item{ID}{character string representing the unique identifier of the
          real employee}
        \item{mh}{integer value representing the man hours assigned}
        \item{mhType}{character string representing the man hour type

           \code{mhType} can be one of the following:
           \describe{
             \item{reg}{working hours not more than 8 hours during a regular
               day}
             \item{rd}{working hours not more than 8 hours during a rest day}
             \item{sh}{working hours not more than 8 hours during a special
               holiday}
             \item{lh}{working hours not more than 8 hours during a legal
               holiday}
             \item{nh}{working hours not more than 8 hours during a
               negotiated holiday}
             \item{rs}{working hours not more than 8 hours during a rest day
               on a special holiday}
             \item{rl}{working hours not more than 8 hours during a rest day
               on a legal holiday}
             \item{rn}{working hours not more than 8 hours during a rest day
               on a negotiated holiday}
             \item{regOT}{working hours more than 8 hours during a regular
               day}
             \item{rdOT}{working hours more than 8 hours during a rest day}
             \item{shOT}{working hours more than 8 hours during a special
               holiday}
             \item{lhOT}{working hours more than 8 hours during a legal
               holiday}
             \item{nhOT}{working hours more than 8 hours during a negotiated
               holiday}
             \item{rsOT}{working hours more than 8 hours during a rest day on
               a special holiday}
             \item{rlOT}{working hours more than 8 hours during a rest day on
               a legal holiday}
             \item{rnOT}{working hours more than 8 hours during a rest day on
               a negotiated holiday}
           }
        }
        \item{month}{integer value representing the month}
        \item{np}{integer value representing the man hours with night premium
          pay}
        \item{costCode}{character string representing accounting cost code
          wherein the man hours is charged}
        \item{sal}{a single character: \code{"a"} or \code{"b"}

          This represents the salary to be applied (see
          \code{\link{payA}} and \code{\link{payB}}).}
        \item{scheme}{character string defining the salary scheme

          This may be \code{"m"} (monthly) or \code{"d"} (daily).}
        \item{isReg}{logical\cr
          Is the employee regular?}
        \item{maxReg}{integer value \cr
          Number of hours the employee is required to report to enjoy full
          salary.}
      }
    \item an \code{\link{Employee-class}} object representing the theoretical
      employee with reduced man hours
    \item an \code{\link{Employee-class}} object representing the real
      employee with reduced man hours
  }
}
\description{
Man hours from an \code{\link{Employee-class}} object representing a real
  employee is assigned to the man hours from an \code{\link{Employee-class}}
  object representing a theoretical employee (manpower requirement). During
  successful assignment, the man hours of both \code{Employee-class} objects
  are reduced.
}
\section{Methods (by class)}{
\itemize{
\item \code{Employee}: Assign \code{reg} hours

\item \code{Staff}: Check for incompatible \code{\link{Staff-class}}
objects

\item \code{Non Staff}: Assign \code{regOT} hours

\item \code{Clerk}: Check for incompatible \code{\link{Clerk-class}}
objects

\item \code{Operation Personnel}: Assign \code{rd}, \code{sh}, \code{lh}, \code{nh},
\code{rs}, \code{rl}, \code{rn}, \code{rdOT}, \code{shOT}, \code{lhOT},
\code{nhOT}, \code{rsOT}, \code{rlPT} and \code{rnOT}

\item \code{Technical}: Check for incompatible \code{\link{Technical-class}}
objects

\item \code{Production Personnel}: Compute for working hours with night premium pay

\item \code{Supervisor}: Check for incompatible \code{\link{Supervisor-class}}
objects

\item \code{Laborer}: Check for incompatible \code{\link{Laborer-class}}
objects

\item \code{Operator}: Check for incompatible \code{\link{Operator-class}}
objects
}}

