% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhDB_fun_assignPrio.R
\name{assignPrio}
\alias{assignPrio}
\title{Assign pool by priority}
\usage{
assignPrio(empReq, empPool, listT, listR)
}
\arguments{
\item{empReq}{passed from \code{\link{getmhDB}}}

\item{empPool}{passed from \code{\link{getmhDB}}}

\item{listT}{list of theoretical employees created from empReq}

\item{listR}{list of real employees created fro empPool}
}
\value{
a list containing the following:
  \enumerate{
    \item man hour database resulting from the assignment
       This is also merged and described well in \code{\link{getmhDB}}.
    \item list of \code{\link{Employee-class}} objects representing the
      theoretical employees with assigned man hours
    \item list of \code{\link{Employee-class}} objects representing the
      real employees with assigned man hours
    \item \code{\link{data.frame}} containing un-assigned man hours of the
      theoretical employees

      This is composed of the following columns:
      \describe{
        \item{month}{integer value representing the month}
        \item{ID}{character string representing the employee requirement}
        \item{mhType}{man hour type}
        \item{mh}{integer value representing the unassigned man hours}
      }
    \item \code{\link{data.frame}} containing un-assigned man hours of the
      real employees

      This is composed of the following columns:
      \describe{
        \item{month}{integer value representing the month}
        \item{ID}{character string representing the employee requirement}
        \item{mhType}{man hour type}
        \item{mh}{integer value representing the unassigned man hours}
      }
    \item man hour database resulting from discarding OT hours. The structure
      is similar to what is being returned by \code{\link{getmhDB}}.
  }
}
\description{
This is used in \code{\link{getmhDB}}.
  Assignment is done in the following order:
  \enumerate{
    \item Any employee with a designated cost code is assigned.
    \item Any regular or probationary employee is assigned.
    \item All employees are assigned.
  }
}
