% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budget.R
\name{budget}
\alias{budget}
\title{Compute budget}
\usage{
budget(xlsxFile, year, forecast = FALSE)
}
\arguments{
\item{xlsxFile}{full path of the spreadsheet}

\item{year}{integer value defining the year to be budgeted}

\item{forecast}{logical value \cr
Compute cost for forecast?}
}
\description{
Given a xlsx file containing the employee pool, requirements, wages, and
  work schedules, the budget is computed.
}
