% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budget.R
\name{budget2}
\alias{budget2}
\title{Compute budget in windows}
\usage{
budget2(myFile, year)
}
\arguments{
\item{myFile}{full path of the spreadsheet}

\item{year}{integer value defining the year to be budgeted}
}
\description{
Given a spreadsheet containing the employee pool, requirements, wages, and
  work schedules, the budget is computed.
}
