% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{capacityEnough}
\alias{capacityEnough}
\title{Is capacity of data.frame enough?}
\usage{
capacityEnough(x, dfLength)
}
\arguments{
\item{x}{\code{data.frame}}

\item{dfLength}{integer representing the number of rows to be appended to x}
}
\value{
\code{TRUE} if number of blank rows is greater than \code{dfLength}.
  \code{FALSE} if number of blank rows is less than or equal to
  \code{dfLength}.
}
\description{
Tests whether a \code{data.frame} has enough blank rows at the bottom.
}
