% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{cleanCC}
\alias{cleanCC}
\title{Clean Cost Center Name}
\usage{
cleanCC(x)
}
\arguments{
\item{x}{character vector}
}
\value{
character vector without leading zero and without non-alpha-numeric
  characters
}
\description{
Removes non-alpha-numeric characters and leading zero in a character vector.
}
