% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{cleanCCS}
\alias{cleanCCS}
\title{Clean Cost Centers}
\usage{
cleanCCS(x)
}
\arguments{
\item{x}{character vector}
}
\value{
character vector without leading zero and without non-alpha-numeric
  characters
}
\description{
Similar to \link{cleanCC} but each element of the vector can be multiple
  cost centers separated by white space.
}
