% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_class.R
\docType{class}
\name{Clerk-class}
\alias{Clerk-class}
\alias{clerk}
\title{An S4 class representing a clerk}
\description{
Clerks are mainly responsible for data encoding. Prior to the creation of
 \code{\link{mansched}}, they were responsible for man hours encoding
 using the spreadsheet template from accounting.
}
\examples{
clerk()
}
\seealso{
\code{\link{NonStaff-class}}
}
