% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_fun_init.R
\name{createEmp}
\alias{createEmp}
\title{Create an employee}
\usage{
createEmp(empClass)
}
\arguments{
\item{empClass}{character string representing the employee sub-class

  Only the end nodes are accepted (see vignette).}
}
\value{
an object that is a sub-class of \code{\link{Employee-class}}
}
\description{
Creates an \code{\link{Employee-class}} object using a switch.
}
