% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_fun.R
\name{dates}
\alias{dates}
\title{Create dates}
\usage{
dates(begin, end)
}
\arguments{
\item{begin}{a character string defining the start date\cr
The accepted format is \code{"yyyy-mm-dd"}.}

\item{end}{a character string defining the end date\cr
The accepted format is \code{"yyyy-mm-dd"}.}
}
\value{
a \code{data.frame} with only one column (\code{"date"})
}
\description{
Creates a \code{\link{data.frame}} of one column containing a range of dates.
}
