% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_class.R
\docType{class}
\name{DepartmentManager-class}
\alias{DepartmentManager-class}
\alias{department_manager}
\title{An S4 class representing a department manager}
\description{
There are two departments existing at Mines Group: Mine Engineering
  Department and Mine Operations Department.
}
\examples{
department_manager()
}
\seealso{
\code{\link{Staff-class}}
}
