% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{dfAppend}
\alias{dfAppend}
\title{Append data.frame}
\usage{
dfAppend(x, y)
}
\arguments{
\item{x}{\code{data.frame} which may contain NA rows at the bottom}

\item{y}{\code{data.frame}}
}
\value{
\code{data.frame} which may contain NA rows at the bottom
}
\description{
Append \code{data.frame} y to \code{data.frame} x.
}
