% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_class.R
\docType{class}
\name{DivisionManager-class}
\alias{DivisionManager-class}
\alias{division_manager}
\title{An S4 class representing a division manager}
\description{
Only one division manager (Mine DivisionManager) is presently employed in
  TMC Mines Group.
}
\examples{
division_manager()
}
\seealso{
\code{\link{Staff-class}}
}
