% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{expandDF}
\alias{expandDF}
\title{Expand \code{data.frame}}
\usage{
expandDF(x)
}
\arguments{
\item{x}{\code{data.frame}}
}
\value{
\code{data.frame} with \code{NA} values at the lower half rows
}
\description{
Expands a \code{data.frame} twice its original number of rows. The expanded
  rows contain only \code{NA} values.
}
