% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_fun_get13mp.R
\name{get13mp}
\alias{get13mp}
\alias{get13mp,Employee-method}
\title{Compute monthly accrued 13th month pay}
\usage{
get13mp(theObject, sal)

\S4method{get13mp}{Employee}(theObject, sal)
}
\arguments{
\item{theObject}{\code{\link{Employee-class}} object}

\item{sal}{a numeric vector of length 2

  This contains 2 monthly salaries of the employee. The first one is salary
  before annual increase while the second one is the adjusted salary.}
}
\value{
a \code{\link{data.frame}} with 12 rows and 3 columns representing
  the accrued 13th month pay per month

  Each row represents a month. The columns are:
  \describe{
     \item{ID}{character string representing the unique identifier of the
       real employee}
     \item{month}{integer value representing the month}
     \item{mp}{numeric value defining the accrued 13th month pay}
  }
}
\description{
13th month pay is pro-rated to the number of months the employee is employed
  for the year. This is then distributed equally to all months of employment.
}
\section{Functions}{
\itemize{
\item \code{get13mp(Employee)}: Compute 13 month pay

}}
