% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_fun.R
\name{getCalDays}
\alias{getCalDays}
\title{Get Calendar Days}
\usage{
getCalDays(cBegin, cEnd = NA, hol, restday)
}
\arguments{
\item{cBegin}{character string representing the start date

  The accepted format is \code{"yyyy-mm-dd"}.}

\item{cEnd}{character string representing the end date

  The accepted format is \code{"yyyy-mm-dd"}.}

\item{hol}{a \code{\link{data.frame}} return by \code{\link{getHol}}}

\item{restday}{character string representing the day of the week defined as
the rest day}
}
\value{
a \code{matrix} with 12 rows and 7 columns

Each row represents a month while each column represents a man hour type.
}
\description{
Using a date range, the available working days type is computed. The
beginning and ending date must be on the same year.
}
