% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_fun_getCBA.R
\name{getCBA}
\alias{getCBA}
\alias{getCBA,Employee-method}
\title{Compute cost for CBA provisions}
\usage{
getCBA(theObject)

\S4method{getCBA}{Employee}(theObject)
}
\arguments{
\item{theObject}{\code{\link{Employee-class}} object}
}
\value{
a \code{\link{data.frame}} with 12 rows and 3 columns representing
  the distributed cost of working long shirts per month

  Each row represents a month. The columns are:
  \describe{
     \item{ID}{character string representing the unique identifier of the
       real employee}
     \item{month}{integer value representing the month}
     \item{benefits}{numeric value defining the cost charged for that month}
  }
}
\description{
Taganito Mine provides items to all regular employees as agreed upon with
  the labor union.
}
\section{Functions}{
\itemize{
\item \code{getCBA,Employee-method}: Compute pro-rated cost for CBA provisions
throughout the year
}}

