% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_fun_getCM.R
\docType{methods}
\name{getCM}
\alias{getCM}
\alias{getCM,Employee-method}
\title{Get cost multiplier}
\usage{
getCM(theObject)

\S4method{getCM}{Employee}(theObject)
}
\arguments{
\item{theObject}{\code{\link{Employee-class}} object}
}
\value{
a \code{\link{data.frame}} with 12 rows and 3 columns

  Each row represents a month while the columns are:
  \describe{
    \item{month}{integer value representing the month}
    \item{ID}{character string representing the unique identifier of the
      employee}
    \item{allow}{numeric value of 0 to 1 representing the cost multiplier}
  }
}
\description{
Based on the period of employment of the personnel, his or her cost
  multipliers throughout the year is computed.
  This multiplier is used for SSS, 13th month pay, and hospital and medical
  expenses.
}
\section{Methods (by class)}{
\itemize{
\item \code{Employee}: Compute allowance multiplier
}}

