% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhDB_fun_getCost.R
\name{getCost}
\alias{getCost}
\title{Compute Personnel Cost}
\usage{
getCost(
  mhDB,
  listR,
  wage,
  forecast = FALSE,
  bonusFactorYearEnd = 1.5,
  absentee = NA,
  monthStart = 1L,
  leaveConversionFactor = 0.5
)
}
\arguments{
\item{mhDB}{a \code{\link{data.frame}} similar to \code{mhDB} return by
\code{\link{getmhDB}}}

\item{listR}{a list of \code{\link{Employee-class}} objects representing the
employee pool}

\item{wage}{a \code{\link{data.frame}} composed of two columns:
  \describe{
    \item{ID}{character string representing a unique identifier of an
      employee}
    \item{s}{integer value representing the salary of an employee

       For rank and file, this value represents the daily salary at the
       beginning of the year. For non-rank and file, this value represents
       the monthly salary at the beginning of the year.}
  }}

\item{forecast}{logical value \cr
Compute cost for forecast?}

\item{bonusFactorYearEnd}{a numeric value for the salary multiplier for the
year-end bonus}

\item{absentee}{a \code{\link{data.frame}} composed of at least two columns:
\describe{
  \item{ID}{character string representing a unique identifier of an
    employee}
  \item{costCenter}{Cost centers wherein the leave commutation or the
    holiday pays are distributed whenever the employee does not report for
    work for a month.
  }
}}

\item{monthStart}{an integer representing the month of start of forecasting}

\item{leaveConversionFactor}{numeric value for the monthly salary fraction
assumed as leave conversion}
}
\value{
a list containing the following:

  \enumerate{
    \item a data.frame with 16 columns

      The last 12 columns represents the cost per month.
      The 1st column is the activity cost center.
      The 2nd column is the description of the general ledger code.
      The 3rd column is the equipment type used.
      The 4th column is the general ledger code.

    \item tabulated total man hours per month per cost center
    \item tabulated 13th month pay per month per cost center
    \item tabulated bonus cost per month per cost center
  }
}
\description{
Calculates monthly personnel cost for each cost center Personnel costs are
  comprised by:
  \describe{
    \item{60300000}{Payroll - Salaries (Regular)}
    \item{60300001}{Payroll - Salaries (Seasonal)}
    \item{60300004}{Payroll - Overtime (Regular)}
    \item{60300005}{Payroll - Overtime (Seasonal)}
    \item{60300008}{Payroll - 13th Month and Other Bonuses (Regular)}
    \item{60300009}{Payroll - 13th Month and Other Bonuses (Seasonal)}
    \item{60300011}{Payroll - Leave Commutation (Regular)}
    \item{60300012}{Payroll - Leave Commutation (Seasonal)}
    \item{60300014}{Payroll - De Minimis Benefits (Regular)}
    \item{60300015}{Payroll - De Minimis Benefits (Seasonal)}
    \item{60300017}{SSS Contributions (Regular)}
    \item{60300018}{SSS Contributions (Seasonal)}
    \item{60300020}{PHIC Contributions (Regular)}
    \item{60300021}{PHIC Contributions (Seasonal)}
    \item{60300023}{HDMF Contributions (Regular)}
    \item{60300024}{HDMF Contributions (Seasonal)}
    \item{60300029}{Other Employee Allowances}
    \item{60300033}{Hospital and Medical Expenses - Employees}
    \item{60300036}{Awards and Prizes}
    \item{60900000}{Life and Medical Insurance - Employees}
    \item{61100000}{Contract Fee - Agency Services}
  }
}
