% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_fun_getGC.R
\docType{methods}
\name{getGC}
\alias{getGC}
\alias{getGC,Employee-method}
\title{Compute gift certificate cost}
\usage{
getGC(theObject)

\S4method{getGC}{Employee}(theObject)
}
\arguments{
\item{theObject}{\code{\link{Employee-class}} object}
}
\value{
a \code{\link{data.frame}} with 12 rows and 3 columns although
  only either the 10th or 12th row may have a value since the cost is charged
  only in October (for seasonal employees) and in December (for regular and
  probationary employees).

  Each row represents a month. The columns are:
  \describe{
     \item{ID}{character string representing the unique identifier of the
       real employee}
     \item{month}{integer value representing the month}
     \item{benefits}{numeric value defining the cost charged for that month}
  }
}
\description{
Gift certificates amounting to 1500 PhP are usually given to employee on
  December for non-seasonal employees. For seasonal employees, the gift
  certificates are awarded in October.
}
\section{Methods (by class)}{
\itemize{
\item \code{Employee}: Summarize gift certificate cost
}}

