% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_fun_getHM.R
\docType{methods}
\name{getHM}
\alias{getHM}
\alias{getHM,Employee-method}
\title{Compute monthly hospital and medical expenses budgeted for employee}
\usage{
getHM(theObject)

\S4method{getHM}{Employee}(theObject)
}
\arguments{
\item{theObject}{\code{\link{Employee-class}} object}
}
\value{
a \code{\link{data.frame}} with 12 rows and 3 columns representing
  the total hospital and medical allowances  receives per month

  Each row represents a month. The columns are:
  \describe{
     \item{ID}{character string representing the unique identifier of the
       real employee}
     \item{month}{integer value representing the month}
     \item{HM}{numeric value defining the total hospital and medical
       allowances budgeted to the employee}
  }
}
\description{
Depending on the employment status of the employee, his or her monthly
  hospital and medical allowances may be the following:
  \enumerate{
    \item 1,600 PhP monthly (reg)
    \item 1,000 PhP at 1st month of employment (pro)
    \item 600 PhP at 1st month of employment (sea)
  }
}
\section{Methods (by class)}{
\itemize{
\item \code{Employee}: Compute hospital and medical expenses
}}

