% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_fun_getHours.R
\docType{methods}
\name{getHours}
\alias{getHours}
\alias{getHours,Employee-method}
\alias{getHours,NonStaff-method}
\alias{getHours,OperationPersonnel-method}
\title{Get remaining working hours}
\usage{
getHours(theObject)

\S4method{getHours}{Employee}(theObject)

\S4method{getHours}{NonStaff}(theObject)

\S4method{getHours}{OperationPersonnel}(theObject)
}
\arguments{
\item{theObject}{an \code{\link{Employee-class}} object}
}
\value{
a \code{\link{data.frame}} representing the remaining man hours

  Each row represents a month while each column represents a man hour type.
}
\description{
This function returns all remaining man hour types including overtime.
}
\section{Methods (by class)}{
\itemize{
\item \code{Employee}: Save \code{reg} hours

\item \code{NonStaff}: Save \code{regOT}

\item \code{OperationPersonnel}: Save all other man hour type
}}

