% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_fun_getHours.R
\name{getHoursL}
\alias{getHoursL}
\title{Get remaining working hours of list of \code{\link{Employee-class}} objects}
\usage{
getHoursL(x)
}
\arguments{
\item{x}{a list of \code{\link{Employee-class}} objects}
}
\value{
an integer value
}
\description{
This function returns the sum of all remaining man hours of a list of
  \code{Employee-class} objects including overtime
}
