% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_fun_getLaborDayShirt.R
\docType{methods}
\name{getLaborDayShirt}
\alias{getLaborDayShirt}
\alias{getLaborDayShirt,Employee-method}
\title{Compute labor day t-shirt cost}
\usage{
getLaborDayShirt(theObject)

\S4method{getLaborDayShirt}{Employee}(theObject)
}
\arguments{
\item{theObject}{\code{\link{Employee-class}} object}
}
\value{
a \code{\link{data.frame}} with 12 rows and 3 columns although
  only the 5th row may have a value since the cost is charged only in May

  Each row represents a month. The columns are:
  \describe{
     \item{ID}{character string representing the unique identifier of the
       real employee}
     \item{month}{integer value representing the month}
     \item{benefits}{numeric value defining the cost charged for that month}
  }
}
\description{
During labor day (May 1), all regular employees are given a shirt worth
  390 Php. This cost is charged in May.
}
\section{Methods (by class)}{
\itemize{
\item \code{Employee}: Summarize shirt cost
}}

