% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyEquipment.R
\name{getLetterCount}
\alias{getLetterCount}
\title{Get letter count}
\usage{
getLetterCount(x)
}
\arguments{
\item{x}{a character string}
}
\value{
a string of at least 26 length having only digits as characters
}
\description{
Gets the frequency per letter of a string
}
