% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_fun_getLongShirt.R
\docType{methods}
\name{getLongShirt}
\alias{getLongShirt}
\alias{getLongShirt,Employee-method}
\title{Compute cost for long shirts issued}
\usage{
getLongShirt(theObject)

\S4method{getLongShirt}{Employee}(theObject)
}
\arguments{
\item{theObject}{\code{\link{Employee-class}} object}
}
\value{
a \code{\link{data.frame}} with 12 rows and 3 columns representing
  the distributed cost of working long shirts per month

  Each row represents a month. The columns are:
  \describe{
     \item{ID}{character string representing the unique identifier of the
       real employee}
     \item{month}{integer value representing the month}
     \item{benefits}{numeric value defining the cost charged for that month}
  }
}
\description{
Taganito Mine issues working long shirts to all in-house employees.
}
\section{Methods (by class)}{
\itemize{
\item \code{Employee}: Compute distributed cost for working long shirts
throughout the year
}}

