% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getMin}
\alias{getMin}
\title{Paired Minimum}
\usage{
getMin(a, b)
}
\arguments{
\item{a}{integer vector}

\item{b}{integer vector}
}
\value{
an integer vector with the least value for each element by element
  comparison between a and b
}
\description{
Get minimum value between two integer vectors
}
