% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_fun_getRB.R
\docType{methods}
\name{getRB}
\alias{getRB}
\alias{getRB,Employee-method}
\title{Compute retention bonus}
\usage{
getRB(theObject)

\S4method{getRB}{Employee}(theObject)
}
\arguments{
\item{theObject}{\code{\link{Employee-class}} object}
}
\value{
a \code{\link{data.frame}} with 12 rows and 3 columns although
  only either the 4th or 10th row may have a value since the cost is charged
  only in April and in October.

  Each row represents a month. The columns are:
  \describe{
     \item{ID}{character string representing the unique identifier of the
       real employee}
     \item{month}{integer value representing the month}
     \item{rb}{numeric value defining the cost charged for that month}
  }
}
\description{
Retention bonus to be given to all seasonal employees.
}
\section{Methods (by class)}{
\itemize{
\item \code{Employee}: Summarize gift certificate cost
}}

