% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_fun_getRiceSub.R
\name{getRiceSub}
\alias{getRiceSub}
\alias{getRiceSub,Employee-method}
\title{Compute the monthly rice subsidy of employee}
\usage{
getRiceSub(theObject)

\S4method{getRiceSub}{Employee}(theObject)
}
\arguments{
\item{theObject}{\code{\link{Employee-class}} object}
}
\value{
a \code{\link{data.frame}} with 12 rows and 3 columns representing
  the rice subsidy the employee receives per month

  Each row represents a month. The columns are:
  \describe{
     \item{ID}{character string representing the unique identifier of the
       real employee}
     \item{month}{integer value representing the month}
     \item{riceSub}{numeric value defining the rice subsidy to be received
       by the employee}
  }
}
\description{
For every month, 1 sack of rice is awarded to the employee.
  Once sack costs around 2,500 PhP.
}
\section{Functions}{
\itemize{
\item \code{getRiceSub,Employee-method}: Compute rice subsidy throughout the year
}}

