% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_fun_getSafetyGadgets.R
\name{getSafetyGadgets}
\alias{getSafetyGadgets}
\alias{getSafetyGadgets,Employee-method}
\title{Compute cost for safety gadgets}
\usage{
getSafetyGadgets(theObject)

\S4method{getSafetyGadgets}{Employee}(theObject)
}
\arguments{
\item{theObject}{\code{\link{Employee-class}} object}
}
\value{
a \code{\link{data.frame}} with 12 rows and 3 columns representing
  the pro-rated cost of safety gadgets per month

  Each row represents a month. The columns are:
  \describe{
     \item{ID}{character string representing the unique identifier of the
       real employee}
     \item{month}{integer value representing the month}
     \item{gadgets}{numeric value defining the cost charged for that month}
  }
}
\description{
Taganito Mine provides steel toe shoes, umbrella, raincoats, and rain boots
  to all in-house employees.
}
\section{Functions}{
\itemize{
\item \code{getSafetyGadgets(Employee)}: Compute price and cost multplier

}}
