% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_fun_getSigningBonus.R
\name{getSigningBonusSea}
\alias{getSigningBonusSea}
\alias{getSigningBonusSea,Employee-method}
\title{Compute Seasonal Employee Signing Bonus}
\usage{
getSigningBonusSea(theObject)

\S4method{getSigningBonusSea}{Employee}(theObject)
}
\arguments{
\item{theObject}{\code{\link{Employee-class}} object}
}
\value{
a \code{\link{data.frame}} with 12 rows and 3 columns.

  Each row represents a month. The columns are:
  \describe{
     \item{ID}{character string representing the unique identifier of the
       real employee}
     \item{month}{integer value representing the month}
     \item{signingBonus}{numeric value defining the cost charged for that
       month}
  }
}
\description{
Signing bonus to be given to all seasonal employees at the start of
  employment.
}
\section{Functions}{
\itemize{
\item \code{getSigningBonusSea(Employee)}: Compute seasonal employee's signing bonus

}}
