% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_class.R
\docType{class}
\name{GroupManager-class}
\alias{GroupManager-class}
\alias{group_manager}
\title{An S4 class representing a group manager}
\description{
No group manager is presently employed in TMC Mines Group. The ex-Mines
  GroupManager was promoted to Mine DivisionManager.
}
\examples{
group_manager()
}
\seealso{
\code{\link{Staff-class}}
}
