% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{holidays}
\alias{holidays}
\title{Holidays in Taganito, Claver in the year 2017}
\format{
\code{\link{data.frame}} with 25 rows and 4 columns:
  \describe{
    \item{Month}{month name}
    \item{Day}{day in the month}
    \item{Type}{type of holiday.}
    \item{Description}{description of holiday}
  }
}
\usage{
holidays
}
\description{
A data-set containing the dates of holidays and their descriptions.
}
\keyword{datasets}
