% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{idAssetGroup}
\alias{idAssetGroup}
\title{Identify asset group}
\usage{
idAssetGroup(x)
}
\arguments{
\item{x}{a character vector representing the equipment}
}
\value{
a character vector representing the asset group
}
\description{
Identifies the asset group of a given equipment from accounting data.
}
