% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyEquipment.R
\name{identifyEquipment}
\alias{identifyEquipment}
\title{Identify equipment group}
\usage{
identifyEquipment(x)
}
\arguments{
\item{x}{a character string representing an equipment}
}
\value{
a character string representing an equipment group
}
\description{
Identifies fleet type of a given equipment.
}
