% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_fun_init.R
\docType{methods}
\name{initEmployee}
\alias{initEmployee}
\alias{initEmployee,Employee-method}
\title{Initialize employee}
\usage{
initEmployee(theObject, ID, name, designation, attendance = 1,
  spareFactor = 1, costCode = "NONE", status = "reg", cBegin, cEnd = NA,
  inHouse, isTheo, restday = "Sunday", hol)

\S4method{initEmployee}{Employee}(theObject, ID, name, designation, attendance,
  spareFactor, costCode, status, isTheo, hol)
}
\arguments{
\item{theObject}{\code{\link{Employee-class}} object}

\item{ID}{character string representing the employee's unique identifier}

\item{name}{character vector representing the name of employee}

\item{designation}{character string representing designation of employee}

\item{attendance}{numeric value used to account for the absenteeism of the
employee}

\item{spareFactor}{numeric value used as manhours multiplier for theoretical
  employees

  For an activity involving continious operation, a spare personnel is
  usually deployed to prevent accidents caused by fatigue. This factor is
  also used to compensate for the absenteeism of employees.}

\item{costCode}{character vector containing the cost codes wherein the
employee will be charged}

\item{status}{character string representing the employment status of the
  employee

  The accepted values are \code{'reg'} (regular), \code{'pro'} (probationary),
  and \code{'sea'} (seasonal).}

\item{isTheo}{logical value showing whether the object is representing a
theoretical employee or not}

\item{hol}{a \code{\link{data.frame}} similar to \code{\link{holidays}}

  This contains all the holidays in the year to be budgeted.}
}
\description{
Initializes employee data and its working hours. This function can be
  applied to both real and theoretical employees. Real employees represent
  the actual manpower pool while theoretical employees represent the required
  manpower and the required man hours.
}
\section{Methods (by class)}{
\itemize{
\item \code{Employee}: Initialize ID, name, designation, attendance,
spareFactor, costCode, status, isTheo
}}

