% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fun_import.R
\name{initEmpReq}
\alias{initEmpReq}
\title{Initialize Employee Requirement}
\usage{
initEmpReq(empReq, sched, hol = NA, year = NA)
}
\arguments{
\item{empReq}{a \code{\link{data.frame}} with 7 columns

  Each row represents an employee requirement for an activity. The columns
  are:
  \describe{
    \item{activity}{character string representing the assigned activity

      Activities listed in this \code{data.frame} must be present in
      \code{workSched}.}
    \item{personnelClass}{character string representing the
      \code{\link{Employee-class}} sub-class to be applied}
    \item{quantity}{integer value representing the number of required
      personnel for the employee requirement}
    \item{spareFactor}{numeric value of at least 1

      This is used as man hours multiplier for deploying a spare personnel.
      For an activity involving continuous operation, a spare personnel is
      usually deployed to prevent accidents caused by fatigue. The
      \code{spareFactor} is also used to compensate for the absenteeism of
      employees.}
    \item{equipment}{character string representing the equipment type to be
      used

      This only applies to \code{\link{Operator-class}} personnel.}
    \item{OT}{integer value defining the number of working hours scheduled
      for more than 8 hours per shift

      This only applies to \code{\link{NonStaff-class}} personnel.}
    \item{costCenter}{character string representing the  cost center
      wherein all personnel costs of this employee requirement will be
      charged}
  }}

\item{sched}{a \code{\link{data.frame}} with 13 columns

  Each row represents one activity. The 1st column is vector of character
  strings representing the name of activity. The 2nd column up to the 13th
  column are all integer vectors representing the number of days an activity
  is scheduled for every month. The column names from the 2nd column up to
  the 13th column are month names. For office activities or for continuous
  activities, their corresponding row may be left blank.}

\item{hol}{a \code{data.frame} similar to \code{\link{holidays}}}

\item{year}{integer value representing the year to be budgeted}
}
\value{
a list of 2:
  \enumerate{
    \item a list of theoretical employees\cr
      Each element of the list is an \code{\link{Employee-class}} object.
      Theoretical employees do not have rest days.
    \item sanitized empReq
  }
}
\description{
Initializes the list of the employee requirements and their corresponding man
  hours.
}
