% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_fun_init.R
\docType{methods}
\name{initREmployee}
\alias{initREmployee}
\alias{initREmployee,Employee-method}
\title{Initialize real employee}
\usage{
initREmployee(theObject, ID, name, designation, attendance = 1,
  costCode = "NONE", status = "reg", cBegin, cEnd = NA, inHouse,
  restday = "Sunday", hol, isRF, equipment)

\S4method{initREmployee}{Employee}(theObject, ID, name, designation, attendance,
  costCode, status, cBegin, cEnd = NA, inHouse, restday)
}
\arguments{
\item{theObject}{\code{\link{Employee-class}} object}

\item{ID}{character string representing the employee's unique identifier}

\item{name}{character vector representing the name of employee}

\item{designation}{character string representing designation of employee}

\item{attendance}{numeric value used to account for the absenteeism of the
employee}

\item{costCode}{character vector containing the cost codes wherein the
employee will be charged}

\item{status}{character string representing the employment status of the
  employee

  The accepted values are \code{"reg"} (regular), \code{"pro"}
  (probationary), and \code{"sea"} (seasonal).}

\item{hol}{a \code{\link{data.frame}} returned by \code{\link{getHol}}

  This contains all the holidays in the year to be budgeted.}

\item{isRF}{logical value \cr
Is the employee rank and file?}

\item{equipment}{character vector representing the equipment types which the
employee is authorized to operate}
}
\description{
Initializes employee data and its working hours. This function can be
  applied only to real employees. Real employees represent the actual
  manpower pool.
}
\section{Methods (by class)}{
\itemize{
\item \code{Employee}: Initialize ID, name, designation, attendance,
spareFactor, costCode, status, isTheo
}}

