% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_fun_init.R
\name{initTEmployee}
\alias{initTEmployee}
\alias{initTEmployee,Employee-method}
\alias{initTEmployee,Staff-method}
\alias{initTEmployee,Clerk-method}
\alias{initTEmployee,OperationPersonnel-method}
\alias{initTEmployee,Operator-method}
\title{Initialize theoretical employee}
\usage{
initTEmployee(
  theObject,
  ID,
  costCenter,
  equipment,
  OT = 0,
  calDays,
  mdtProb,
  spareFactor = 1,
  monthSched = NA
)

\S4method{initTEmployee}{Employee}(theObject, ID, costCenter, spareFactor = 1)

\S4method{initTEmployee}{Staff}(
  theObject,
  ID,
  costCenter,
  calDays,
  mdtProb,
  spareFactor = 1,
  monthSched = NA
)

\S4method{initTEmployee}{Clerk}(
  theObject,
  ID,
  costCenter,
  OT = 0,
  calDays,
  mdtProb,
  spareFactor = 1,
  monthSched = NA
)

\S4method{initTEmployee}{OperationPersonnel}(
  theObject,
  ID,
  costCenter,
  OT = 0,
  mdtProb,
  spareFactor = 1,
  monthSched = NA
)

\S4method{initTEmployee}{Operator}(
  theObject,
  ID,
  costCenter,
  equipment,
  OT = 0,
  mdtProb,
  spareFactor = 1,
  monthSched = NA
)
}
\arguments{
\item{theObject}{\code{\link{Employee-class}} object}

\item{ID}{character string representing the employee's unique identifier}

\item{costCenter}{character string representing the cost center wherein the
employee's man hour cost will be charged}

\item{equipment}{character string representing the equipment type which the
employee was given authority to operate}

\item{OT}{integer value defining the budgeted overtime hours}

\item{calDays}{a \code{\link{data.frame}} returned by
\code{\link{getCalDays}}}

\item{mdtProb}{a \code{data.frame} returned by \code{\link{getMDTProb}}}

\item{spareFactor}{numeric value of greater than or equal to 1 used as man
hours multiplier}

\item{monthSched}{integer vector of length 12

  This represents the number of days an activity is scheduled for each month.
  This argument is used only for \code{\link{ProductionPersonnel-class}}
  objects.}
}
\value{
\code{\link{Employee-class}} object
}
\description{
Initialize employee data and its working hours. This function can be applied
  only to theoretical employees. Theoretical employees represent the manpower
  requirement for a scheduled activity.
}
\section{Functions}{
\itemize{
\item \code{initTEmployee(Employee)}: Initialize ID, costCenter and spareFactor

\item \code{initTEmployee(Staff)}: Initialize reg

\item \code{initTEmployee(Clerk)}: Initialize reg and regOT

\item \code{initTEmployee(OperationPersonnel)}: Initialize all man hour type using man day type
probabilities

\item \code{initTEmployee(Operator)}: Initialize equipment

}}
