% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_fun_isReg.R
\name{isReg}
\alias{isReg}
\alias{isReg,Employee-method}
\title{isReg}
\usage{
isReg(theObject)

\S4method{isReg}{Employee}(theObject)
}
\arguments{
\item{theObject}{an \code{\link{Employee-class}} object}
}
\value{
logical value
}
\description{
Is the employee regular?
}
\section{Functions}{
\itemize{
\item \code{isReg,Employee-method}: Throws an error if employment status is not valid.
}}

