% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_fun_isRF.R
\name{isRF}
\alias{isRF}
\alias{isRF,Employee-method}
\alias{isRF,Staff-method}
\alias{isRF,NonStaff-method}
\alias{isRF,Clerk-method}
\alias{isRF,Technical-method}
\alias{isRF,Supervisor-method}
\alias{isRF,Laborer-method}
\alias{isRF,Operator-method}
\title{isRF}
\usage{
isRF(theObject)

\S4method{isRF}{Employee}(theObject)

\S4method{isRF}{Staff}(theObject)

\S4method{isRF}{NonStaff}(theObject)

\S4method{isRF}{Clerk}(theObject)

\S4method{isRF}{Technical}(theObject)

\S4method{isRF}{Supervisor}(theObject)

\S4method{isRF}{Laborer}(theObject)

\S4method{isRF}{Operator}(theObject)
}
\arguments{
\item{theObject}{an \code{\link{Employee-class}} object}
}
\value{
logical value
}
\description{
Is the employee rank and file?
}
\section{Functions}{
\itemize{
\item \code{isRF,Employee-method}: Set \code{NA} as default value

\item \code{isRF,Staff-method}: All staff are not rank and file

\item \code{isRF,NonStaff-method}: All \code{\link{NonStaff-class}} objects have isRF slot

\item \code{isRF,Clerk-method}: Clerks may or may not be rank and file

\item \code{isRF,Technical-method}: Technical personnel are not rank and file

\item \code{isRF,Supervisor-method}: Supervisors are not rank and file

\item \code{isRF,Laborer-method}: Laborers are rank and file

\item \code{isRF,Operator-method}: Operators are rank and file
}}

