% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyEquipment.R
\name{isSubset}
\alias{isSubset}
\title{Is the string a subset?}
\usage{
isSubset(x, y)
}
\arguments{
\item{x}{a character string to be tested}

\item{y}{a string return by \code{\link{getLetterCount}}

  \code{y} comes from a string wherein the string \code{x} might be a subset.}
}
\value{
logical value
}
\description{
Test whether the string \code{x} is a subset of another string.
}
