% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_class.R
\docType{class}
\name{Laborer-class}
\alias{Laborer-class}
\alias{laborer}
\title{An S4 class representing a laborer}
\description{
Samplers, checkers, spotters, plumbers and \emph{etc.} are classified under
  this class.
}
\examples{
laborer()
}
\seealso{
\code{\link{ProductionPersonnel-class}}
}
