% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{mansched-package}
\alias{mansched}
\alias{mansched-package}
\title{Tools for generating man power cost for budgeting and forecasting in
  Taganito Mine}
\description{
This packages contains \code{R} functions that can automate man power
  assignments and computation of personnel costs. This can be helpful during
  forecasting and budgeting.
}
\details{
Budget season in Taganito Mine means sleepless nights for its personnel,
 especially the personnel from the Mine Engineering Department. Man hours
 budgeting is a very tiresome task. This job requires estimating work
 schedules, manpower requirements and lastly, manual encoding of
 man hours into the spreadsheet template.

The spreadsheet template is full of details and its size is stupendous. It
 has at least 12 sheets. Each sheet has more than a thousand rows and the
 number of columns are magnitudes more than the number of rows.

Encoding the man hours into the template requires great amount of patience
 and time. Also, finding encoding mistakes is also a discouraging task. This
 package is an attempt to automate the task of manually encoding the man
 hours data for manpower cost computation.
}
\author{
\strong{Maintainer}: Basil Eric Rabi \email{basil.rabi@tmc.nickelasia.com}

}
