% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{mansched2-package}
\alias{mansched2}
\alias{mansched2-package}
\title{Tools for generating man hours database budget for TMC Mines Group}
\description{
This packages contains \code{R} functions that may help the computation of
  man hours of TMC Mines Group personnel during budget season.
}
\details{
Budget season in Taganito Mine means sleepless nights for its personnel,
 especially the personnel from the Mine Engineering Department. Man hours
 budgeting is a very tiresome task. This job requires estimating work
 schedules, manpower requirements and lastly, manual encoding of
 man hours into the spreadsheet template.

The spreadsheet template is full of details and its size is stupendous. It
 has at least 12 sheets. Each sheet has more than a thousand rows and the
 number of columns are magnitudes more than the number of rows.

Encoding the man hours into the template requires great amount of patience
 and time. Also, finding encoding mistakes is also a discouraging task. This
 package is an attempt to automate the task of manually encoding the man
 hours data for manpower cost computation.
}
\author{
\strong{Maintainer}: Basil Eric Rabi \email{basil.rabi@tmc.nickelasia.com}

}
