% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_fun_mergeEmp.R
\name{mergeEmp}
\alias{mergeEmp}
\title{Merge theoretical employee list}
\usage{
mergeEmp(x)
}
\arguments{
\item{x}{a list of \code{\link{Employee-class}} representing theoretical
  employees with the same class and the same costCenter slot

  For \code{\link{Operator-class}}, the equipment slot must also be the same.}
}
\value{
an \code{\link{Employee-class}} object
}
\description{
Combines all man hour types of a list of theoretical employees.
  All man hours are stored in only one theoretical employee.
}
