% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_class.R
\docType{class}
\name{NonStaff-class}
\alias{NonStaff-class}
\alias{non_staff}
\title{An S4 class representing an employee who is not a staff}
\description{
This class is entitled to premium pay for overtime work.
  \code{NonStaff-class} is further classified into \code{\link{Clerk-class}}
  and \code{\link{OperationPersonnel-class}}.
}
\section{Slots}{

\describe{
\item{\code{isRF}}{logical value \cr
Is the employee rank and file?}

\item{\code{regOT}}{integer vector with length of 12

  This represents the number of regular overtime hours budgeted per month.}
}}

\examples{
non_staff()
}
\seealso{
\code{\link{Employee-class}}
}
