% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_fun_normEmp.R
\name{normEmp}
\alias{normEmp}
\alias{normEmp,Employee-method}
\alias{normEmp,NonStaff-method}
\alias{normEmp,OperationPersonnel-method}
\alias{normEmp,Operator-method}
\title{Strip all over time hours from an employee}
\usage{
normEmp(theObject)

\S4method{normEmp}{Employee}(theObject)

\S4method{normEmp}{NonStaff}(theObject)

\S4method{normEmp}{OperationPersonnel}(theObject)

\S4method{normEmp}{Operator}(theObject)
}
\arguments{
\item{theObject}{an \code{\link{Employee-class}} object}
}
\value{
an \code{\link{Employee-class}} object with no over time man hour
  type and a dummy cost code.
}
\description{
This function returns an \code{\link{Employee-class}} object with over time
  hours removed (see \code{\link{premium}}). The cost code is also changed
  into a dummy cost code \code{"0-0"}.
}
\section{Functions}{
\itemize{
\item \code{normEmp,Employee-method}: Assign dummy cost code

\item \code{normEmp,NonStaff-method}: Assign zero to regOT

\item \code{normEmp,OperationPersonnel-method}: Assign zero to \code{rd}, \code{sh}, \code{lh},
\code{rs}, \code{rl}, \code{rn}, \code{rdOT}, \code{shOT}, \code{lhOT},
\code{nhOT}, \code{rsOT}, \code{rlOT}, \code{rnOT}. If the employee is
regular, also assign zero to \code{nh}.

\item \code{normEmp,Operator-method}: Take only one equipment
}}

