% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_class.R
\docType{class}
\name{OperationPersonnel-class}
\alias{OperationPersonnel-class}
\alias{operation_personnel}
\title{An S4 class representing an operation personnel}
\description{
Operation-related personnel are classified under this class. These personnel
have different scheduled rest days and thus have a different method of
calculating premium for overtime work compared to \code{\link{Clerk-class}}.
This class is further classified into \code{\link{Technical-class}} and
\code{\link{ProductionPersonnel-class}}.
}
\section{Slots}{

\describe{
\item{\code{rd}}{integer vector with length of 12

  This represents budgeted working hours of not more than 8 hours per day
  rendered on a scheduled rest day. Each element of the vector represents a
  month.}

\item{\code{rdOT}}{integer vector with length of 12

  This represents budgeted working hours of more than 8 hours per day
  rendered on a scheduled rest day. Each element of the vector represents a
  month.}

\item{\code{sh}}{integer vector with length of 12

  This represents budgeted working hours of not more than 8 hours per day
  rendered on a special holiday. Each element of the vector represents a
  month.

  Only regular employees are entitled to a premium pay for this type of
  holiday.}

\item{\code{shOT}}{integer vector with length of 12

  This represents budgeted working hours of more than 8 hours per day
  rendered on a special holiday. Each element of the vector represents a
  month.

  Only regular employees are entitled to a premium pay for this type of
  holiday.}

\item{\code{lh}}{integer vector with length of 12

  This represents budgeted working hours of not more than 8 hours per day
  rendered on a legal holiday. Each element of the vector represents a
  month.}

\item{\code{lhOT}}{integer vector with length of 12

  This represents budgeted working hours of more than 8 hours per day
  rendered on a legal holiday. Each element of the vector represents a
  month.}

\item{\code{nh}}{integer vector with length of 12

  This represents budgeted working hours of not more than 8 hours per day
  rendered on a negotiated holiday. Each element of the vector represents a
  month.

  Only regular employees are entitled to a premium pay for this type of
  holiday.}

\item{\code{nhOT}}{integer vector with length of 12

  This represents budgeted working hours of more than 8 hours per day
  rendered on a negotiated holiday. Each element of the vector represents a
  month.

  Only regular employees are entitled to a premium pay for this type of
  holiday.}

\item{\code{rs}}{integer vector with length of 12

  This represents budgeted working hours of not more than 8 hours per day
  rendered on a special holiday and a rest day. Each element of the vector
  represents a month.

  Only regular employees are entitled to a compounded premium pay for this
  type of holiday.}

\item{\code{rsOT}}{integer vector with length of 12

  This represents budgeted working hours of more than 8 hours per day
  rendered on a special holiday and a rest day. Each element of the vector
  represents a month.

  Only regular employees are entitled to a compounded premium pay for this
  type of holiday.}

\item{\code{rl}}{integer vector with length of 12

  This represents budgeted working hours of not more than 8 hours per day
  rendered on a legal holiday and a rest day. Each element of the vector
  represents a month.}

\item{\code{rlOT}}{integer vector with length of 12

  This represents budgeted working hours of more than 8 hours per day
  rendered on a legal holiday and a rest day. Each element of the vector
  represents a month.}

\item{\code{rn}}{integer vector with length of 12

  This represents budgeted working hours of not more than 8 hours per day
  rendered on a negotiated holiday and a rest day. Each element of the vector
  represents a month.

  Only regular employees are entitled to a compounded premium pay for this
  type of holiday.}

\item{\code{rnOT}}{integer vector with length of 12

  This represents budgeted working hours of more than 8 hours per day
  rendered on a negotiated holiday and a rest day. Each element of the vector
  represents a month.

  Only regular employees are entitled to a compounded premium pay for this
  type of holiday.}
}}

\examples{
operation_personnel()
}
\seealso{
\code{\link{NonStaff-class}}
}
