% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_class.R
\docType{class}
\name{Operator-class}
\alias{Operator-class}
\alias{operator}
\title{An S4 class representing an operator}
\description{
This class includes authorized heavy and light equipment operators and
  drivers for trucks and services vehicles.
}
\section{Slots}{

\describe{
\item{\code{equipment}}{character vector containing equipment types which the
employee is authorized to operate}
}}

\examples{
operator()
}
\seealso{
\code{\link{Production Personnel-class}}
}
