% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{payA}
\alias{payA}
\title{Payment scheme for regular non-RF}
\format{
\code{\link{data.frame}} with 12 rows and two columns
}
\usage{
payA
}
\description{
Salary increase of a regular non-RF employee is effective on July 1.
}
\keyword{datasets}
