% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{payB}
\alias{payB}
\title{Payment scheme for regular RF}
\format{
\code{\link{data.frame}} with 12 rows and two columns
}
\usage{
payB
}
\description{
Salary increase of a regular RF employee is effective on February 1
}
\keyword{datasets}
