% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PHIC}
\alias{PHIC}
\title{Employer's PhilHealth Contribution}
\format{
An object of class \code{data.frame} with 28 rows and 3 columns.
}
\usage{
PHIC
}
\description{
Provided by accounting during 2020 forecasting.
}
\keyword{datasets}
