% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{premium.nonRF}
\alias{premium.nonRF}
\title{Wage premiums for non-Rank and File}
\format{
An object of class \code{data.frame} with 16 rows and 7 columns.
}
\usage{
premium.nonRF
}
\description{
Wage premiums for non-Rank and File
}
\seealso{
\code{\link{premium}} \code{\link{premium.RF}}
}
\keyword{datasets}
