% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{premium}
\alias{premium}
\title{Wage premiums}
\format{\code{\link{data.frame}} with 16 rows and columns

  Each row represents the premium pay for a man hour type. The columns are
  cosisted by:
  \describe{
    \item{type}{man hour type (see \code{\link{assignEmp}})}
    \item{premiumR}{wage premium enjoyed by regular employees}
    \item{premiumS}{wage premium enjoyed by non-regular employees}
    \item{npR}{night premium enjoyed by regular employees}
    \item{npS}{night premium enjoyed by non-regular employees}
  }}
\usage{
premium
}
\description{
Hourly salary multipliers according to man hour types along with the thier
  corresponding night premiums.
}
\keyword{datasets}
