% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{premium.RF}
\alias{premium.RF}
\title{Wage premiums for Rank and File}
\format{
An object of class \code{data.frame} with 16 rows and 7 columns.
}
\usage{
premium.RF
}
\description{
Wage premiums for Rank and File
}
\seealso{
\code{\link{premium}} \code{\link{premium.nonRF}}
}
\keyword{datasets}
