% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_class.R
\docType{class}
\name{ProductionPersonnel-class}
\alias{ProductionPersonnel-class}
\alias{production_personnel}
\title{An S4 class representing a production personnel}
\description{
Production-related personnel are classified under this class. They may be
  assigned during night shift, thus, they are entitled for a night premium
  pay. This class is further classified into \code{\link{Supervisor-class}},
  \code{\link{Laborer-class}} and \code{\link{Operator-class}}.
}
\examples{
production_personnel()
}
\seealso{
\code{\link{OperationPersonnel-class}}
}
