% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{rmLead0}
\alias{rmLead0}
\title{Remove leading zero}
\usage{
rmLead0(x)
}
\arguments{
\item{x}{character vector}
}
\value{
character vector with leading zero removed
}
\description{
Remove leading zeroes for purely numeric characters.
}
