% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{rmS}
\alias{rmS}
\title{Remove space}
\usage{
rmS(x)
}
\arguments{
\item{x}{character vector}
}
\value{
character vector without 2 adjacent space characters
}
\description{
Spaces character is removed from all elements of the character vector.
}
