% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{rmWS}
\alias{rmWS}
\title{Remove white space}
\usage{
rmWS(x)
}
\arguments{
\item{x}{character vector}
}
\value{
character vector without 2 adjacent space characters
}
\description{
Spaces with more than 2-character length are removed and are replaced wth 1
  space. Trailing and leading spaces are also removed.
}
