% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sheet_empPool.R
\name{sanityCheckEmpPool}
\alias{sanityCheckEmpPool}
\title{Sanity Check for Employee Pool Sheet}
\usage{
sanityCheckEmpPool(empPool)
}
\arguments{
\item{empPool}{see \link{initEmpPool}}
}
\value{
empPool with the columns cBegin and cEnd converted to characters.
}
\description{
Sanity Check for Employee Pool Sheet
}
