% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sheet_empReq.R
\name{sanityCheckEmpReq}
\alias{sanityCheckEmpReq}
\title{Sanity Check for Employee Requirement Sheet}
\usage{
sanityCheckEmpReq(empReq)
}
\arguments{
\item{empReq}{see \link{initEmpReq}}
}
\value{
empReq with the the following changes:
  \enumerate{
    \item \code{spareFactor} converted to numeric
    \item \code{quantity} and \code{OT} changed to integer
    \item \code{NA} and0 quantities removed
  }
}
\description{
Sanity Check for Employee Requirement Sheet
}
