% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sheet_hol.R
\name{sanityCheckHol}
\alias{sanityCheckHol}
\title{Sanity Check for Holiday Sheet}
\usage{
sanityCheckHol(hol)
}
\arguments{
\item{hol}{see \link{getmhDB}}
}
\value{
hol with \code{Day} converted to integer
}
\description{
Sanity Check for Holiday Sheet
}
