% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sheet_wage.R
\name{sanityCheckWage}
\alias{sanityCheckWage}
\title{Sanity Check for Wage Sheet}
\usage{
sanityCheckWage(wage, empPool)
}
\arguments{
\item{wage}{see \link{getCost}}

\item{empPool}{see \link{initEmpPool}}
}
\value{
wage with
}
\description{
Sanity Check for Wage Sheet
}
