% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_class.R
\docType{class}
\name{Section Head-class}
\alias{Section Head-class}
\alias{section_head}
\title{An S4 class representing a section head}
\description{
Foremen and department officers-in-charge are classified as section head.
  Some of the sections in Mines Group are Materials Handling, Survey, Grade
  Control, Conveyor System, Shipment and Road Maintenance.
}
\examples{
section_head()
}
\seealso{
\code{\link{Staff-class}}
}
