% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SSS}
\alias{SSS}
\title{Employer's SSS Contribution}
\format{An object of class \code{data.frame} with 31 rows and 3 columns.}
\usage{
SSS
}
\description{
Details are posted in
\url{https://www.sss.gov.ph/sss/appmanager/pages.jsp?page=scheduleofcontribution}.
}
\keyword{datasets}
