% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_class.R
\docType{class}
\name{Staff-class}
\alias{Staff-class}
\alias{staff}
\title{An S4 class representing a staff}
\description{
These are employees that are classified as
  \code{\link{Division Manager-class}},
  \code{\link{Group Manager-class}},
  \code{\link{Department Manager-class}}
  \code{\link{Section Head-class}}. In TMC, department managers and up are
  termed "Senior Staff" while section heads are termed "Junior Staff".
}
\examples{
staff()
}
\seealso{
\code{\link{Employee-class}}
}
