% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_class.R
\docType{class}
\name{Supervisor-class}
\alias{Supervisor-class}
\alias{supervisor}
\title{An S4 class representing a supervisor}
\description{
This class may include field engineers and lead-men. Supervisors report to
  their respective foremen or shift-in-charge. They are responsible for
  monitoring their \code{\link{Laborer-class}} and
  \code{\link{Operator-class}} personnel.
}
\examples{
supervisor()
}
\seealso{
\code{\link{ProductionPersonnel-class}}
}
