% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_class.R
\docType{class}
\name{Technical-class}
\alias{Technical-class}
\alias{technical}
\title{An S4 class representing a technical personnel}
\description{
Engineers, some draftsmen and assistants providing technical support are
  classified under this S4 class. These personnel report on duty during the
  day only.
}
\examples{
technical()
}
\seealso{
\code{\link{OperationPersonnel-class}}
}
