% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_functions.R
\name{testAndExpand}
\alias{testAndExpand}
\title{Test and Expand \code{data.frame}}
\usage{
testAndExpand(x, dfLength)
}
\arguments{
\item{x}{\code{data.frame}}

\item{dfLength}{integer representing the number of rows to be appended to x}
}
\value{
\code{data.frame} with NA rows at the bottom
}
\description{
Tests a \code{data.frame} whether there is enough blank rows for appending.
  If there is not enough slot, the \code{data.frame} is expanded with blank
  rows.
}
