% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{validEmpClass}
\alias{validEmpClass}
\title{Valid \code{personnelClass}}
\format{character vector}
\usage{
validEmpClass
}
\description{
A character vector containing the valid personnel classes used in this
  package.
}
\keyword{datasets}
