% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{validEmpStatus}
\alias{validEmpStatus}
\title{Valid employment \code{status}}
\format{
character vector
}
\usage{
validEmpStatus
}
\description{
A character vector containing the valid employment status in Taganito Mine.
}
\keyword{datasets}
