% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{validEquipment}
\alias{validEquipment}
\title{Valid equipment types}
\format{
character vector:

  \describe{
    \item{BREAKER}{Breaker}
    \item{CRANE}{crane}
    \item{CT}{bull dozer}
    \item{DT}{dump truck}
    \item{FL}{fuel truck}
    \item{FORKLIFT}{forklift}
    \item{MDT}{Mini Dump Truck}
    \item{MOC}{Mobile Crusher}
    \item{MOS}{Mobile Screen}
    \item{PB}{Pump boat}
    \item{PMT}{prime mover (trailer)}
    \item{RADIO}{Radio}
    \item{RG}{road grader}
    \item{SB}{service bus}
    \item{SP}{service pickup}
    \item{ST}{service truck}
    \item{TX}{tracked excavator}
    \item{VC}{vibrating compactor}
    \item{WL}{wheeled pay loader}
    \item{WTL}{water truck}
    \item{WX}{wheeled excavator}
  }
}
\usage{
validEquipment
}
\description{
A character vector containing the valid equipment types in Taganito Mine.
}
\keyword{datasets}
