% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acSAP}
\alias{acSAP}
\title{Cost accounting codes in SAP}
\format{
\code{\link{data.frame}} with 778 rows and 2 columns:
  \describe{
    \item{code}{accounting code}
    \item{description}{cost component}
  }
}
\usage{
acSAP
}
\description{
SAP cost codes used by accounting.
}
\keyword{datasets}
