% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhDB_fun_assignPool.R
\name{assignPool}
\alias{assignPool}
\title{Assign pool to requirement}
\usage{
assignPool(empReq, empPool, listT, listR, prioStat = NA,
  prioCode = FALSE)
}
\arguments{
\item{empReq}{passed from \code{\link{assignPrio}}}

\item{empPool}{passed from \code{\link{assignPrio}}}

\item{listT}{list of theoretical employees created from empReq}

\item{listR}{list of real employees created fro empPool}

\item{prioStat}{character vector defining the employee status that are
prioritized in assigning man hours}

\item{prioCode}{logical value \cr
Is cost code prioritized in assigning?}
}
\value{
a list containing the following:
  \enumerate{
    \item remaining empReq
    \item remaining empPool
    \item remaining listT
    \item remaining listR
    \item man hour database resulting from the assignment

      This is also merged and described well in \code{\link{getmhDB}}.
  }
}
\description{
This is used in \code{\link{assignPrio}}.
}
