% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budget.R
\name{budget}
\alias{budget}
\title{Compute budget}
\usage{
budget(myFile, year)
}
\arguments{
\item{myFile}{full path of the spreadsheet}

\item{year}{integer value defining the year to be budgeted}
}
\description{
Given a xlsx file containing the employee pool, requirements, wages, and
  work schedules, the budget is computed.
}
