% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_class.R
\docType{class}
\name{Employee-class}
\alias{Employee-class}
\alias{employee}
\title{An S4 class defining an employee}
\description{
This class is based on the organizational structure ofTaganito Mining
  Corporation's (TMC) Mines Group. This can be used in representing
  both real and theoretical employees.
}
\details{
Real employees are actual persons represented as \code{Employee-class}
  objects. Theoretical employees in this context are employee requirements
  represented as \code{Employee-class} objects.

This class is further classified into \code{\link{Staff-class}} and
  \code{\link{Non Staff-class}}.
}
\section{Slots}{

\describe{
\item{\code{ID}}{character string defining the employee's unique identifier}

\item{\code{name}}{character string defining the name of employee}

\item{\code{designation}}{character string defining the designation of the
employee}

\item{\code{attendance}}{numeric value estimating the attendance rate of the
  employee

  This defaults to 1 for theoretical employees. For real employees,
  a value of not more than 1 but greater than 0.5 may be used.}

\item{\code{spareFactor}}{numeric value estimating the man power spare factor for
  an activity

  This is usually used as man hours multiplier when a spare operator is
  assigned. The spare factor defaults to 1 for real employees. For
  theoretical employees, a value greater than 1 may be used.}

\item{\code{costCode}}{character vector containing the cost codes wherein the
employee will be charged}

\item{\code{status}}{character string defining the employment status of the employee}

\item{\code{cBegin}}{character string defining the date of start of employment of the
employee}

\item{\code{cEnd}}{character string defining the date of end of employment of the
employee}

\item{\code{inHouse}}{logical value \cr
Is the employee's accomodation given by the company?}

\item{\code{restday}}{character string defining the day of the week wherein the
employee is not required to report to work}

\item{\code{holHours}}{integer vector with length of 12

  This represents the total number hours consisting of only the first eight
  hours of a holiday not scheduled on a rest day of employee. Each element
  of the vector represents a month.

  Even if the employee will not report on duty, he or she will be paid by
  this number of hours.
  This slot only applies to real employees.}

\item{\code{totHours}}{integer vector with length of 12

  This represents the number of hours per month which is equated to the
  employee's basic monthly wage.}

\item{\code{maxReg}}{integer vector with length of 12

  This represents the number of hours wherein the employee must report to
  work.}

\item{\code{reg}}{integer vector with length of 12

  This represents the number of regular hours budgeted per month.}
}}

\examples{
employee()
}
