% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_fun_getAllowance.R
\name{getAllowance}
\alias{getAllowance}
\title{Compute the monthly allowance of employee}
\usage{
getAllowance(theObject)
}
\arguments{
\item{theObject}{\code{\link{Employee-class}} object}
}
\value{
a \code{\link{data.frame}} with 12 rows and 3 columns representing
  the total allowances the employee receives per month

  Each row represents a month. The columns are:
  \describe{
     \item{ID}{character string representing the unique identifier of the
       real employee}
     \item{month}{integer value representing the month}
     \item{cost}{numeric value defining the total allowances to be received
       by the employee}
  }
}
\description{
Depending on the attributes of the employee, his or her monthly allowances
  may include the following:
  \enumerate{
    \item food
    \item lighting
    \item housing
  }
}
