% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhDB_fun_getCost.R
\name{getCost}
\alias{getCost}
\title{Compute Personnel Cost}
\usage{
getCost(mhDB, listR, wage, forecast = FALSE, bonusFactorYearEnd = 1.5)
}
\arguments{
\item{mhDB}{a \code{\link{data.frame}} similar to \code{mhDB} return by
\code{\link{getmhDB}}}

\item{listR}{a list of \code{\link{Employee-class}} objects representing the
employee pool}

\item{wage}{a \code{\link{data.frame}} composed of two columns:
  \describe{
    \item{ID}{character string representing a unique identifier of an
      employee}
    \item{s}{integer value representing the salary of an employee

       For rank and file, this value represents the daily salary at the
       beginning of the year. For non-rank and file, this value represents
       the monthly salary at the beginning of the year.}
  }}

\item{forecast}{logical value \cr
Compute cost for forecast?}

\item{bonusFactorYearEnd}{a numeric value for the salary multiplier for the
year-end bonus}
}
\value{
a list containing the following:

  \enumerate{
    \item a data.frame with 16 columns

      The last 12 columns represents the cost per month. The first column
      is a concatenated string from the 2nd column (Cost Center) and 4th
      column (accounting code). The 3rd column is the description of the
      accounting code.

    \item tabulated total man hours per month per cost code
    \item tabulated 13th month pay per month per cost code
    \item tabulated bonus cost per month per cost code
    \item a data.frame with 15 columns

      This is simlar to the first object returned but composed of only
      seasonal employee-related costs.
      Also, there is no concat column since this will not be linked to the
      budget templates.
  }
}
\description{
Calculates monthly personnel cost for each cost code. Personnel costs are
  comprised by:
  \describe{
    \item{521001}{Salaries-Regular}
    \item{521021}{OT Pay - Regular}
    \item{521002}{Salaries-Seasonal}
    \item{521022}{OT Pay - Seasonal}
    \item{521004}{Employees Allowance}
    \item{521005}{Employee Benefits}
    \item{521006}{Premium SSS, EC}
    \item{521007}{Prem-HDMF (Pag-ibig)}
    \item{521008}{Philhealth}
    \item{521012}{Leave Commutation}
    \item{521011}{Food Allowance / Rice Subsidy}
    \item{521017}{Hospital and Medical Expenses}
    \item{521009}{13th Month Pay}
    \item{521018}{HMO}
    \item{524037}{Grouplife}
    \item{523011}{Safety Gadgets}
    \item{522010}{CF Manpower Services}
  }
}
