% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_fun_getGroupLife.R
\name{getGroupLife}
\alias{getGroupLife}
\alias{getGroupLife,Employee-method}
\alias{getGroupLife,Staff-method}
\alias{getGroupLife,DivisionManager-method}
\alias{getGroupLife,GroupManager-method}
\alias{getGroupLife,DepartmentManager-method}
\alias{getGroupLife,SectionHead-method}
\alias{getGroupLife,NonStaff-method}
\title{Compute the monthly cost for group life insurance premium}
\usage{
getGroupLife(theObject)

\S4method{getGroupLife}{Employee}(theObject)

\S4method{getGroupLife}{Staff}(theObject)

\S4method{getGroupLife}{DivisionManager}(theObject)

\S4method{getGroupLife}{GroupManager}(theObject)

\S4method{getGroupLife}{DepartmentManager}(theObject)

\S4method{getGroupLife}{SectionHead}(theObject)

\S4method{getGroupLife}{NonStaff}(theObject)
}
\arguments{
\item{theObject}{\code{\link{Employee-class}} object}
}
\value{
a \code{\link{data.frame}} with 12 rows and 3 columns representing
  the total HMO costs budgeted for the employee per month

  Each row represents a month. The columns are:
  \describe{
     \item{ID}{character string representing the unique identifier of the
       real employee}
     \item{month}{integer value representing the month}
     \item{gl}{numeric value defining the total costs budgeted to the
       employee per month}
  }
}
\description{
Life insurance privilage is only available to regular employees. The higher
  the job grade of the employee, the higher the privilage enjoyed.
}
\section{Functions}{
\itemize{
\item \code{getGroupLife,Employee-method}: Compute cost multiplier

\item \code{getGroupLife,Staff-method}: Compute premium list

\item \code{getGroupLife,DivisionManager-method}: Compute monthly cost

\item \code{getGroupLife,GroupManager-method}: Compute monthly cost

\item \code{getGroupLife,DepartmentManager-method}: Compute monthly cost

\item \code{getGroupLife,SectionHead-method}: Compute monthly cost

\item \code{getGroupLife,NonStaff-method}: Compute monthly cost
}}

