% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_fun_getHMO.R
\docType{methods}
\name{getHMO}
\alias{getHMO}
\alias{getHMO,Employee-method}
\alias{getHMO,Staff-method}
\alias{getHMO,NonStaff-method}
\title{Compute the monthly cost for health and maintenance organization}
\usage{
getHMO(theObject)

\S4method{getHMO}{Employee}(theObject)

\S4method{getHMO}{Staff}(theObject)

\S4method{getHMO}{NonStaff}(theObject)
}
\arguments{
\item{theObject}{\code{\link{Employee-class}} object}
}
\value{
a \code{\link{data.frame}} with 12 rows and 3 columns representing
  the total HMO costs budgeted for the employee per month

  Each row represents a month. The columns are:
  \describe{
     \item{ID}{character string representing the unique identifier of the
       real employee}
     \item{month}{integer value representing the month}
     \item{hmo}{numeric value defining the total costs budgeted to the
       employee per month}
  }
}
\description{
HMO privilage is only available to regular employees. The higher the job
  grade of the employee, the higher the privilage per dependent.
}
\section{Methods (by class)}{
\itemize{
\item \code{Employee}: Compute cost multiplier

\item \code{Staff}: Compute premium list

\item \code{NonStaff}: Compute monthly cost
}}

