% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getHours}
\alias{getHours}
\title{Get remaining working hours}
\usage{
getHours(employee)
}
\arguments{
\item{employee}{an \code{\link{Employee-class}} object}
}
\value{
a \code{\link{data.frame}} representing the remaining man hours

  Each row represents a month while each column represents a man hour type.
}
\description{
This function returns all remaining man hour types including overtime.
}
