% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_fun_getHours.R
\docType{methods}
\name{getHours}
\alias{getHours}
\alias{getHours,Employee-method}
\alias{getHours,Non Staff-method}
\alias{getHours,Operation Personnel-method}
\title{Get remaining working hours}
\usage{
getHours(theObject)

\S4method{getHours}{Employee}(theObject)

\S4method{getHours}{`Non Staff`}(theObject)

\S4method{getHours}{`Operation Personnel`}(theObject)
}
\arguments{
\item{theObject}{an \code{\link{Employee-class}} object}
}
\value{
a 12 by 16 \code{\link{matrix}} representing the remaining man hours


  Each row represents a month while each column represents a man hour type.
}
\description{
This function returns all remaining man hour types including overtime.
}
\section{Methods (by class)}{
\itemize{
\item \code{Employee}: Save \code{reg} hours

\item \code{Non Staff}: Save \code{regOT}

\item \code{Operation Personnel}: Save all other man hour type
}}

