% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_fun_getSigningBonus.R
\name{getSigningBonus}
\alias{getSigningBonus}
\alias{getSigningBonus,Employee-method}
\alias{getSigningBonus,Staff-method}
\alias{getSigningBonus,NonStaff-method}
\title{Compute RF Signing Bonus}
\usage{
getSigningBonus(theObject)

\S4method{getSigningBonus}{Employee}(theObject)

\S4method{getSigningBonus}{Staff}(theObject)

\S4method{getSigningBonus}{NonStaff}(theObject)
}
\arguments{
\item{theObject}{\code{\link{Employee-class}} object}
}
\value{
a \code{\link{data.frame}} with 12 rows and 3 columns although
  only one row may have a value since the cost is charged only in one month.

  Each row represents a month. The columns are:
  \describe{
     \item{ID}{character string representing the unique identifier of the
       real employee}
     \item{month}{integer value representing the month}
     \item{signingBonus}{numeric value defining the cost charged for that
       month}
  }
}
\description{
Signing bonus is given to Rank-and-File employees at the year of the start
  of a new collective bargaining agreement.
}
\section{Functions}{
\itemize{
\item \code{getSigningBonus,Employee-method}: Compute cost multiplier

\item \code{getSigningBonus,Staff-method}: No signing bonus for Staff

\item \code{getSigningBonus,NonStaff-method}: Compute monthly cost
}}

