% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fun_import.R
\name{initEmpPool}
\alias{initEmpPool}
\title{Initialize Employee Pool}
\usage{
initEmpPool(empPool, hol = NA, year = NA, forecast = FALSE)
}
\arguments{
\item{empPool}{a \code{data.frame} with 87 columns

  Each row represents a real employee. The columns are:
  \describe{
    \item{ID}{character string representing the employee's unique identifier}
    \item{name}{character string representing the employee's name}
    \item{designation}{character string representing the employee's
      designation}
    \item{personnelClass}{character string representing the
      \code{\link{Employee-class}} sub-class to be applied}
    \item{field}{logical value \cr
      Is the employee always in the field?}
    \item{equipment}{character string representing the equipment types which
      the employee was authorized to operate.

      Each equipment type must be separated by a space.}
    \item{costCenter}{character string representing the preferred cost center
      wherein the employee shall be assigned.

      Multiple cost centers must be separated by spaces.}
    \item{status}{character string representing the employment status of the
      employee

      Accepted values are \code{"reg"} (regular), \code{"pro"}
      (probationary), \code{'sea'} (seasonal), and \code{'age'} (agency).}
    \item{cBegin}{character string representing the date of employment of
      personnel

      The accepted format is ISO 8601 which expresses a day as
      \code{"yyyy-mm-dd"}.}
    \item{cEnd}{character string representing the date of termination
      employment of the seasonal employee

      The accepted format is ISO 8601 which expresses a day as
      \code{"yyyy-mm-dd"}.}
    \item{inHouse}{logical value \cr
      Is the employee's accommodation given by the company?}
    \item{restday}{character string representing the day of the week wherein
      the employee is not required to report to work}
    \item{isRF}{logical value \cr
      Is the employee rank and file?}
    \item{d.rd_1}{integer value defining how many rest days the employee can
      report to work on January}
    \item{d.rd_2}{integer value defining how many rest days the employee can
      report to work on February}
    \item{d.rd_3}{integer value defining how many rest days the employee can
      report to work on March}
    \item{d.rd_4}{integer value defining how many rest days the employee can
      report to work on April}
    \item{d.rd_5}{integer value defining how many rest days the employee can
      report to work on May}
    \item{d.rd_6}{integer value defining how many rest days the employee can
      report to work on June}
    \item{d.rd_7}{integer value defining how many rest days the employee can
      report to work on July}
    \item{d.rd_8}{integer value defining how many rest days the employee can
      report to work on August}
    \item{d.rd_9}{integer value defining how many rest days the employee can
      report to work on September}
    \item{d.rd_10}{integer value defining how many rest days the employee can
      report to work on October}
    \item{d.rd_11}{integer value defining how many rest days the employee can
      report to work on November}
    \item{d.rd_12}{integer value defining how many rest days the employee can
      report to work on December}
    \item{d.ho_1}{integer value defining how many holidays the employee can
      report to work on January}
    \item{d.ho_2}{integer value defining how many holidays the employee can
      report to work on February}
    \item{d.ho_3}{integer value defining how many holidays the employee can
      report to work on March}
    \item{d.ho_4}{integer value defining how many holidays the employee can
      report to work on April}
    \item{d.ho_5}{integer value defining how many holidays the employee can
      report to work on May}
    \item{d.ho_6}{integer value defining how many holidays the employee can
      report to work on June}
    \item{d.ho_7}{integer value defining how many holidays the employee can
      report to work on July}
    \item{d.ho_8}{integer value defining how many holidays the employee can
      report to work on August}
    \item{d.ho_9}{integer value defining how many holidays the employee can
      report to work on September}
    \item{d.ho_10}{integer value defining how many holidays the employee can
      report to work on October}
    \item{d.ho_11}{integer value defining how many holidays the employee can
      report to work on November}
    \item{d.ho_12}{integer value defining how many holidays the employee can
      report to work on December}
    \item{d.rh_1}{integer value defining how many rest days on holidays the
      employee can report to work on January}
    \item{d.rh_2}{integer value defining how many rest days on holidays the
      employee can report to work on February}
    \item{d.rh_3}{integer value defining how many rest days on holidays the
      employee can report to work on March}
    \item{d.rh_4}{integer value defining how many rest days on holidays the
      employee can report to work on April}
    \item{d.rh_5}{integer value defining how many rest days on holidays the
      employee can report to work on May}
    \item{d.rh_6}{integer value defining how many rest days on holidays the
      employee can report to work on June}
    \item{d.rh_7}{integer value defining how many rest days on holidays the
      employee can report to work on July}
    \item{d.rh_8}{integer value defining how many rest days on holidays the
      employee can report to work on August}
    \item{d.rh_9}{integer value defining how many rest days on holidays the
      employee can report to work on September}
    \item{d.rh_10}{integer value defining how many rest days on holidays the
      employee can report to work on October}
    \item{d.rh_11}{integer value defining how many rest days on holidays the
      employee can report to work on November}
    \item{d.rh_12}{integer value defining how many rest days on holidays the
      employee can report to work on December}
    \item{dcc_1}{character string representing the cost center wherein the
      idle hours of the employee will be charged on January}
    \item{dcc_2}{character string representing the cost center wherein the
      idle hours of the employee will be charged on February}
    \item{dcc_3}{character string representing the cost center wherein the
      idle hours of the employee will be charged on March}
    \item{dcc_4}{character string representing the cost center wherein the
      idle hours of the employee will be charged on April}
    \item{dcc_5}{character string representing the cost center wherein the
      idle hours of the employee will be charged on May}
    \item{dcc_6}{character string representing the cost center wherein the
      idle hours of the employee will be charged on June}
    \item{dcc_7}{character string representing the cost center wherein the
      idle hours of the employee will be charged on July}
    \item{dcc_8}{character string representing the cost center wherein the
      idle hours of the employee will be charged on August}
    \item{dcc_9}{character string representing the cost center wherein the
      idle hours of the employee will be charged on September}
    \item{dcc_10}{character string representing the cost center wherein the
      idle hours of the employee will be charged on October}
    \item{dcc_11}{character string representing the cost center wherein the
      idle hours of the employee will be charged on November}
    \item{dcc_12}{character string representing the cost center wherein the
      idle hours of the employee will be charged on December}
    \item{JAN}{number of dependents for the month}
    \item{FEB}{number of dependents for the month}
    \item{MAR}{number of dependents for the month}
    \item{APR}{number of dependents for the month}
    \item{MAY}{number of dependents for the month}
    \item{JUN}{number of dependents for the month}
    \item{JUL}{number of dependents for the month}
    \item{AUG}{number of dependents for the month}
    \item{SEP}{number of dependents for the month}
    \item{OCT}{number of dependents for the month}
    \item{NOV}{number of dependents for the month}
    \item{DEC}{number of dependents for the month}
    \item{VL}{numeric value representing the number of vacation leave
      credits}
    \item{SL}{numeric value representing the number of sick leave
      credits}
    \item{a_1}{numeric value representing the attendance rate of
      employee on January}
    \item{a_2}{numeric value representing the attendance rate of
      employee on February}
    \item{a_3}{numeric value representing the attendance rate of
      employee on March}
    \item{a_4}{numeric value representing the attendance rate of
      employee on April}
    \item{a_5}{numeric value representing the attendance rate of
      employee on May}
    \item{a_6}{numeric value representing the attendance rate of
      employee on June}
    \item{a_7}{numeric value representing the attendance rate of
      employee on July}
    \item{a_8}{numeric value representing the attendance rate of
      employee on August}
    \item{a_9}{numeric value representing the attendance rate of
      employee on September}
    \item{a_10}{numeric value representing the attendance rate of
      employee on October}
    \item{a_11}{numeric value representing the attendance rate of
      employee on November}
    \item{a_12}{numeric value representing the attendance rate of
      employee on December}
  }}

\item{hol}{a \code{data.frame} similar to \code{\link{holidays}}}

\item{year}{integer value representing the year to be budgeted}

\item{forecast}{logical value \cr
Compute cost for forecast?}
}
\value{
list of of 2:
  \enumerate{
    \item a list of real employees\cr
      Each element of the list is an \code{\link{Employee-class}} object.
    \item sanitized empPool
  }
}
\description{
Initializes the list of the pool of employees and their corresponding
  maximum man hours which can be assigned to any activity.
}
