% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fun_import.R
\name{initEmpPool}
\alias{initEmpPool}
\title{Initialize Employee Pool}
\usage{
initEmpPool(empPool, hol = NA, year = NA)
}
\arguments{
\item{empPool}{a \code{data.frame} with 13 columns

  Each row represents a real employee. The columns are:
  \describe{
    \item{ID}{character string representing the employee's unique identifier}
    \item{name}{character string representing the employee's name}
    \item{designation}{character string representing the employee's
      designation}
    \item{personnelClass}{character string representing the
      \code{\link{Employee-class}} sub-class to be applied}
    \item{attendance}{numeric value representing the attendance rate of
      employee

      This value must be less than or equal to one and more than 0.5.}
    \item{equipment}{character string representing the equipment types which
      the employee was authorized to operate.

      Each equipment type must be separated by a space.}
    \item{costCode}{character string representing the preferred cost code
      wherein the employee shall be assigned.

      Multiple cost codes must be separated by spaces.}
    \item{status}{character string representing the employment status of the
      employee

      Accepted values are \code{"reg"} (regular), \code{"pro"}
      (probationary), and \code{'sea'} (seasonal).}
    \item{cBegin}{character string representing the date of employment of
      personnel

      The accepted format is ISO 8601 which expresses a day as
      \code{"yyyy-mm-dd"}.}
    \item{cEnd}{character string representing the date of termination
      employment of the seasonal employee

      The accepted format is ISO 8601 which expresses a day as
      \code{"yyyy-mm-dd"}.}
    \item{inHouse}{logical value \cr
      Is the employee's accommodation given by the company?}
    \item{restday}{character string representing the day of the week wherein
      the employee is not required to report to work}
    \item{isRF}{logical value \cr
      Is the employee rank and file?}
  }}

\item{hol}{a \code{data.frame} similar to \code{\link{holidays}}}

\item{year}{integer value representing the year to be budgeted}
}
\value{
list of real employees

  Each element of the list is an \code{\link{Employee-class}} object.
}
\description{
Initializes the list of the pool of employees and their corresponding
  maximum man hours which can be assigned to any activity.
}
