% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_fun_init.R
\docType{methods}
\name{initREmployee}
\alias{initREmployee}
\alias{initREmployee,Employee-method}
\alias{initREmployee,Staff-method}
\alias{initREmployee,NonStaff-method}
\alias{initREmployee,Clerk-method}
\alias{initREmployee,OperationPersonnel-method}
\alias{initREmployee,Technical-method}
\alias{initREmployee,Supervisor-method}
\alias{initREmployee,Laborer-method}
\alias{initREmployee,Operator-method}
\title{Initialize real employee}
\usage{
initREmployee(theObject, ID, name, designation, attendance = 1,
  costCode = "NONE", status = "reg", cBegin, cEnd = NA,
  inHouse = FALSE, restday = "Sunday", hol, RF = FALSE, equipment,
  OT = 3, d.rd = NA, d.ho = NA, d.rh = NA, dcc = "NA",
  forecast = FALSE)

\S4method{initREmployee}{Employee}(theObject, ID, name, designation,
  attendance = 1, costCode = "NONE", status = "reg", cBegin,
  cEnd = NA, inHouse = FALSE, restday = "Sunday", hol, d.rd = NA,
  d.ho = NA, d.rh = NA, dcc = "NA", forecast = FALSE)

\S4method{initREmployee}{Staff}(theObject, ID, name, designation,
  attendance = 1, costCode = "NONE", status = "reg", cBegin,
  cEnd = NA, inHouse = FALSE, restday = "Sunday", hol, d.rd = NA,
  d.ho = NA, d.rh = NA, dcc = "NA", forecast = FALSE)

\S4method{initREmployee}{NonStaff}(theObject, ID, name, designation,
  attendance = 1, costCode = "NONE", status = "reg", cBegin,
  cEnd = NA, inHouse = FALSE, restday = "Sunday", hol, OT = 3,
  d.rd = NA, d.ho = NA, d.rh = NA, dcc = "NA", forecast = FALSE)

\S4method{initREmployee}{Clerk}(theObject, ID, name, designation,
  attendance = 1, costCode = "NONE", status = "reg", cBegin,
  cEnd = NA, inHouse = FALSE, restday = "Sunday", hol, RF = FALSE,
  OT = 3, d.rd = NA, d.ho = NA, d.rh = NA, dcc = "NA",
  forecast = FALSE)

\S4method{initREmployee}{OperationPersonnel}(theObject, ID, name,
  designation, attendance = 1, costCode = "NONE", status = "reg",
  cBegin, cEnd = NA, inHouse = FALSE, restday = "Sunday", hol,
  OT = 3, d.rd = NA, d.ho = NA, d.rh = NA, dcc = "NA",
  forecast = FALSE)

\S4method{initREmployee}{Technical}(theObject, ID, name, designation,
  attendance = 1, costCode = "NONE", status = "reg", cBegin,
  cEnd = NA, inHouse = FALSE, restday = "Sunday", hol, OT = 3,
  d.rd = NA, d.ho = NA, d.rh = NA, dcc = "NA", forecast = FALSE)

\S4method{initREmployee}{Supervisor}(theObject, ID, name, designation,
  attendance = 1, costCode = "NONE", status = "reg", cBegin,
  cEnd = NA, inHouse = FALSE, restday = "Sunday", hol, OT = 3,
  d.rd = NA, d.ho = NA, d.rh = NA, dcc = "NA", forecast = FALSE)

\S4method{initREmployee}{Laborer}(theObject, ID, name, designation,
  attendance = 1, costCode = "NONE", status = "reg", cBegin,
  cEnd = NA, inHouse = FALSE, restday = "Sunday", hol, OT = 3,
  d.rd = NA, d.ho = NA, d.rh = NA, dcc = "NA", forecast = FALSE)

\S4method{initREmployee}{Operator}(theObject, ID, name, designation,
  attendance = 1, costCode = "NONE", status = "reg", cBegin,
  cEnd = NA, inHouse = FALSE, restday = "Sunday", hol, equipment,
  OT = 3, d.rd = NA, d.ho = NA, d.rh = NA, dcc = "NA",
  forecast = FALSE)
}
\arguments{
\item{theObject}{\code{\link{Employee-class}} object}

\item{ID}{character string representing the employee's unique identifier}

\item{name}{character vector representing the name of employee}

\item{designation}{character string representing designation of employee}

\item{attendance}{numeric value used to account for the absenteeism of the
  employee

  This may be used as correction factor to estimate the workable man hours of
  an employee.}

\item{costCode}{character vector containing the cost codes wherein the
employee will be charged}

\item{status}{character string representing the employment status of the
  employee

  The accepted values are \code{"reg"} (regular), \code{"pro"}
  (probationary), and \code{"sea"} (seasonal).}

\item{cBegin}{character string defining the date wherein the employment
  contract of the employee began

  The accepted format is \code{"yyyy-mm-dd"}.}

\item{cEnd}{character string defining the date wherein the employment
  contract of the employee will end

  The accepted format is \code{"yyyy-mm-dd"}.}

\item{inHouse}{logical value \cr
Is the employee's accommodation given by the company?}

\item{restday}{character string defining the day of the week wherein the
employee is not required to report to work}

\item{hol}{a \code{\link{data.frame}} returned by \code{\link{getHol}}

  This contains all the holidays in the year to be budgeted.}

\item{RF}{logical value \cr
Is the employee rank and file?}

\item{equipment}{character string representing the equipment types which the
  employee was given authority to operate

  Equipment types are separated by spaces and must be in upper case.}

\item{OT}{integer value defining the budgeted overtime hours}

\item{d.rd}{integer value defining the number of rest days the employee
is budgeted to report to work per month}

\item{d.ho}{integer value defining the number of exclusive holidays the
employee is budgeted to report to work per month}

\item{d.rh}{integer value defining the number of rest days during a holiday
the employee is budgeted to report to work per month}

\item{dcc}{character string defining the cost code wherein the excess regular
man hours of the employee will be charged}

\item{forecast}{logical value \cr
Compute cost for forecast?}
}
\value{
an \code{\link{Employee-class}} object
}
\description{
Initializes employee data and its working hours. This function can be
  applied only to real employees. Real employees represent the actual
  manpower pool.
}
\section{Methods (by class)}{
\itemize{
\item \code{Employee}: Initialize ID, name, designation, attendance,
costCode, status, cBegin, cEnd, inHouse, restday and calDays
(see \code{\link{getCalDays}})

\item \code{Staff}: Return only the \code{\link{Staff-class}} object

\item \code{NonStaff}: Initialize regOT

\item \code{Clerk}: Initialize RF and return only the
\code{\link{Clerk-class}} object

\item \code{OperationPersonnel}: Initialize rd, rdOT, sh, shOT, lh, lhOT, nh, nhOT,
rs, rsOT, rl, rlOT, rn, rnOT. Return only the
\code{\link{OperationPersonnel-class}} object.

\item \code{Technical}: Return only the \code{\link{Technical-class}}
object. All \code{Technical-class} personnel are not rank and file.

\item \code{Supervisor}: Return only the
\code{\link{Supervisor-class}} object. All \code{Supervisor-class}
employees are not rank and file

\item \code{Laborer}: Return only the
\code{\link{Laborer-class}} object. All \code{Laborer-class}
employees are rank and file

\item \code{Operator}: Initialize equipment

\code{\link{Operator-class}} personnel may be authorized to operate multiple
  types of equipment. In Taganito Mining Corporation, some of the equipment
  types are listed in \code{\link{validEquipment}}.

  All \code{Operator-class} employees are rank and file.
}}

