% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{isRF}
\alias{isRF}
\title{isRF}
\usage{
isRF(theObject)
}
\arguments{
\item{theObject}{an \code{\link{Employee-class}} object}
}
\value{
logical value
}
\description{
Is the employee rank and file? Also checks whether an RF flag is correctly
assigned to the employees.
}
