% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{normEmp}
\alias{normEmp}
\title{Strip all over time hours from an employee}
\usage{
normEmp(emp)
}
\arguments{
\item{emp}{an \code{\link{Employee-class}} object}
}
\value{
a \code{\link{data.frame}} of dropped man-hours. The columns of the
  which is similar to what is being returned in \code{\link{getmhDB}}.
}
\description{
This function modifies an \code{\link{Employee-class}} object with overtime
  hours removed (see \code{\link{premium}}). The cost center is also changed
  into a dummy cost center \code{"0-0"}.
}
