% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PHIC}
\alias{PHIC}
\title{Employer's PHIC Contribution}
\format{An object of class \code{data.frame} with 29 rows and 3 columns.}
\usage{
PHIC
}
\description{
Details are posted in
\url{https://www.philhealth.gov.ph/partners/employers/contri_tbl.html}.
}
\keyword{datasets}
