% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{premium}
\alias{premium}
\title{Wage premiums}
\format{\code{\link{data.frame}} with 16 rows and 7 columns

  Each row represents the premium pay for a man hour type. The columns are
  cosisted by:
  \describe{
    \item{mhType}{character value representing the man hour type (see
      \code{\link{assignEmp}})}
    \item{premiumR}{numeric value representing the wage premium enjoyed by
      regular employees}
    \item{premiumS}{numeric value representing wage premium enjoyed by
      non-regular employees}
    \item{npR}{numeric value representing night premium enjoyed by regular
      employees}
    \item{npS}{numeric value representing night premium enjoyed by
      non-regular employees}
    \item{isOT.R}{logical value \cr
      Is man hour type of regular employee treated as overtime work?}
    \item{isOT.S}{logical value \cr
      Is man hour type of non-regular employee treated as overtime work?}

  }}
\usage{
premium
}
\description{
Hourly salary multipliers according to man hour types along with the their
  corresponding night premiums.
}
\seealso{
\code{\link{premium.RF}} \code{\link{premium.nonRF}}
}
\keyword{datasets}
