% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{validEquipment}
\alias{validEquipment}
\title{Valid equipment types}
\format{character vector:

  \describe{
    \item{CRANE}{crane}
    \item{CT}{bull dozer}
    \item{DT}{dump truck}
    \item{FL}{fuel truck}
    \item{FORKLIFT}{forklift}
    \item{MOC}{Mobile Crusher}
    \item{MOS}{Mobile Screen}
    \item{RG}{road grader}
    \item{SB}{service bus}
    \item{SP}{service pickup}
    \item{TT}{prime mover (trailer truck)}
    \item{TX}{tracked excavator}
    \item{VC}{vibrating compactor}
    \item{WL}{wheeled payloader}
    \item{WTL}{water truck}
    \item{WX}{wheeled excavator}
  }}
\usage{
validEquipment
}
\description{
A character vector containing the valid equipment types in Taganito Mine.
}
\keyword{datasets}
